projeto-rec/back/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
joblib==0.12.1
MarkupSafe==1.0
numpy==1.15.0
scikit-surprise==1.0.6
scipy==1.1.0
six==1.11.0
Werkzeug==0.14.1
